% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGuiPlots.R
\name{makeGuiPlots}
\alias{makeGuiPlots}
\title{Make plots for the rCTM model}
\usage{
makeGuiPlots(modelOutput, inputParms)
}
\arguments{
\item{modelOutput}{A list of three data frames. Contains output of the function "runMemWithCohorts".}

\item{inputParms}{A list of model parameters passed from the GUI input.}
}
\value{
a list of 6 plots similar to those in MEM 3.4
}
\description{
This function delivers a set of standard plots for the Marsh Equilibrium Cohort model.
}
